# [pkg/translate/promremotewrite] downscale exponential histograms for prometheus