**Description:** 

Implement down-scaling of exponential histograms to Prometheus native histograms in Prometheus remote writer.
Configuration of down-scaling is TBD.

**Link to tracking Issue:**

Fixes: #17565 

**Testing:**

Unit tests.

**Documentation:**

TBD