@krajorama 2023-07-09
Current benchmark compared to original:
```
goos: linux
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/prometheusremotewrite
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                            │ bench-orig.txt │            bench-new.txt            │
                            │     sec/op     │   sec/op     vs base                │
ConvertBucketLayout/gap_0-8      17.30µ ± 1%   20.88µ ± 1%  +20.70% (p=0.000 n=10)
ConvertBucketLayout/gap_1-8      16.05µ ± 1%   19.09µ ± 0%  +18.89% (p=0.000 n=10)
ConvertBucketLayout/gap_2-8      15.61µ ± 1%   18.68µ ± 1%  +19.69% (p=0.000 n=10)
ConvertBucketLayout/gap_3-8      18.58µ ± 1%   21.51µ ± 2%  +15.77% (p=0.000 n=10)
geomean                          16.85µ        20.01µ       +18.75%
```