@krajorama 2023-07-09
After some optimizations I got this benchmark compared to baseline:
```
goos: linux
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/prometheusremotewrite
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                            │ bench-orig.txt │           bench-new3.txt           │
                            │     sec/op     │   sec/op     vs base               │
ConvertBucketLayout/gap_0-8      17.30µ ± 1%   18.18µ ± 1%  +5.07% (p=0.000 n=10)
ConvertBucketLayout/gap_1-8      16.05µ ± 1%   16.91µ ± 1%  +5.34% (p=0.000 n=10)
ConvertBucketLayout/gap_2-8      15.61µ ± 1%   16.94µ ± 1%  +8.55% (p=0.000 n=10)
ConvertBucketLayout/gap_3-8      18.58µ ± 1%   19.47µ ± 1%  +4.76% (p=0.000 n=10)
geomean                          16.85µ        17.84µ       +5.92%
```

Looking at the profile, half the time is spent in reallocating the deltas and probably spans slices. That could be something to look at , although for small numbers it's probably fine.