@Dogacel 2023-07-17
Hoping this will resolve our issues with prometheus histograms 🙏 Putting the error we have seen here for visibility.

> Exporting failed. The error is not retryable. Dropping data. {"kind": "exporter", "data_type": "metrics", "name": "prometheusremotewrite", "error": "Permanent error: cannot convert exponential to native histogram. Scale must be <= 8 and >= -4, was 20;